requirements.txt
streamlit
plotly
requests
neo4j
networkx
pandas
python-dotenv
